feature(KB-147): add ArtistMapper to art service functionality, remove setFieldsToDto method, modify step of catologue number

+++ src/main/java/com/dmitryshundrik/knowledgebase/model/music/Composition.java
@@ -29,7 +29,7 @@ public class Composition extends AbstractEntity {
     private String title;
 
     @Column(name = "CATALOG_NUMBER")
-    private Integer catalogNumber;
+    private Double catalogNumber;
 
     @ManyToOne
     private Musician musician;
